Extract WebhookPayloadFactory to a payload factories module
