+++ prlint/github/tests/test_payload_factories/test_webhook_payload_factory.py
@@ -1,24 +1,28 @@
 from django.test import TestCase
 
-from ...factories import WebhookPayloadFactory
+from ..payload_factories import PingPayloadFactory
 
 
 class TestWebhookPayloadFactory(TestCase):
 
     def test_happy(self):
         """
-        WebhookPayloadFactory creates a default payload
+        PingPayloadFactory creates a default payload
         """
-        result = WebhookPayloadFactory()
+        result = PingPayloadFactory()
 
         self.assertIsInstance(result, dict)
         self.assertGreater(result['zen'], '')
         self.assertIsInstance(result['hook_id'], int)
         self.assertGreater(result['hook_id'], 0)
         self.assertEqual(result['hook']['events'], ['pull_request'])
+        # self.assertGreater(result['repository']['id'], 999)
 
     def test_custom(self):
-        result = WebhookPayloadFactory(
+        """
+        PingPayloadFactory can generate custom events
+        """
+        result = PingPayloadFactory(
             hook_events=['a', 'b', 'c']
         )
